src/recommendation-agent/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
requests==2.28.1
google-generativeai==0.3.1
